avb/milan: move aecp-aem-types.h to common
