+++ src/modules/module-avb/aecp-aem.h
@@ -8,7 +8,7 @@
 #define AVB_AEM_H
 
 #include "common/aecp.h"
-#include "aecp-cmd-resp/aecp-aem-types.h"
+#include "common/aecp-aem-types.h"
 #include "common/aecp-aem-counters.h"
 
 struct avb_packet_aecp_aem_acquire {
