+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-cmd-set-sampling-rate.c
@@ -6,7 +6,7 @@
 #include "../common/aecp-aem-state.h"
 #include "../common/descriptors.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../common/aecp-aem-types.h"
 #include "aecp-aem-unsol-helper.h"
 #include "aecp-aem-cmd-set-sampling-rate.h"
