+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-configuration.c
@@ -6,7 +6,7 @@
 #include "../common/aecp-aem-state.h"
 #include "../common/aecp.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../common/aecp-aem-types.h"
 
 #include "../common/aecp-aem-descriptors.h"
 #include "aecp-aem-configuration.h"
