+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-descriptors.c
@@ -8,7 +8,7 @@
 #include "../common/aecp-aem-state.h"
 #include "../common/descriptors.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../common/aecp-aem-types.h"
 
 #include "aecp-aem-descriptors.h"
