+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-get-avb-info.c
@@ -6,7 +6,7 @@
 
 #include "../common/descriptors.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../common/aecp-aem-types.h"
 
 #include "aecp-aem-descriptors.h"
