+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-lock-entity.c
@@ -9,7 +9,7 @@
 #include "../common/aecp-aem-state.h"
 #include "../common/descriptors.h"
 
-#include "aecp-aem-types.h"
+#include "../common/aecp-aem-types.h"
 #include "aecp-aem-lock-entity.h"
 #include "aecp-aem-helpers.h"
 #include "aecp-aem-unsol-helper.h"
