+++ src/modules/module-avb/aecp-cmd-resp/aecp-aem-unsol-notifications.c
@@ -6,7 +6,7 @@
 #include "../common/aecp-aem-state.h"
 #include "../common/descriptors.h"
 
-#include "aecp-aem-types.h"
+#include "../common/aecp-aem-types.h"
 #include "aecp-aem-unsol-notifications.h"
 #include "aecp-aem-helpers.h"
