+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-cmd-get-name.c
@@ -9,7 +9,7 @@
 #include "../../module-avb/common/aecp-aem-controls.h"
 
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 #include "aecp-aem-cmd-resp-common.h"
 #include "aecp-aem-cmd-get-name.h"
