+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-cmd-set-clock-source.c
@@ -6,7 +6,7 @@
 #include "../../module-avb/common/aecp-aem-descriptors.h"
 #include "../../module-avb/common/aecp-aem-state.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 
 #include "aecp-aem-cmd-set-clock-source.h"
 #include "aecp-aem-unsol-helper.h"
