+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-cmd-set-name.c
@@ -7,7 +7,7 @@
 #include "../../module-avb/common/aecp-aem-state.h"
 #include "../../module-avb/common/descriptors.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 
 #include "aecp-aem-cmd-set-name.h"
 #include "aecp-aem-name-common.h"
