+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-cmd-set-stream-format.c
@@ -1,7 +1,7 @@
 #include "../../module-avb/common/aecp-aem-state.h"
 #include "../../module-avb/common/descriptors.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 #include "aecp-aem-unsol-helper.h"
 
 #include "aecp-aem-cmd-set-stream-format.h"
