+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-configuration.c
@@ -6,7 +6,7 @@
 #include "../../module-avb/common/aecp-aem-state.h"
 #include "../../module-avb/common/aecp.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 
 #include "../../module-avb/common/aecp-aem-descriptors.h"
 #include "aecp-aem-configuration.h"
