+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-get-avb-info.c
@@ -6,7 +6,7 @@
 
 #include "../../module-avb/common/descriptors.h"
 #include "aecp-aem-helpers.h"
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 
 #include "aecp-aem-descriptors.h"
