+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-lock-entity.c
@@ -9,7 +9,7 @@
 #include "../../module-avb/common/aecp-aem-state.h"
 #include "../../module-avb/common/descriptors.h"
 
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 #include "aecp-aem-lock-entity.h"
 #include "aecp-aem-helpers.h"
 #include "aecp-aem-unsol-helper.h"
