+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-unsol-helper.h
@@ -11,7 +11,7 @@
 #include "../../module-avb/common/aecp-aem-state.h"
 #include "../../module-avb/common/aecp-aem.h"
 #include "../../module-avb/common/aecp.h"
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 
 #define AECP_AEM_MIN_PACKET_LENGTH AVB_PACKET_MIN_SIZE
 /**
