+++ src/modules/module-milan/aecp-cmd-resp/aecp-aem-unsol-notifications.c
@@ -6,7 +6,7 @@
 #include "../../module-avb/common/aecp-aem-state.h"
 #include "../../module-avb/common/descriptors.h"
 
-#include "aecp-aem-types.h"
+#include "../../module-avb/common/aecp-aem-types.h"
 #include "aecp-aem-unsol-notifications.h"
 #include "aecp-aem-helpers.h"
